Refactor backend code
