+++ digiprod_gen/frontend/session.py
@@ -2,15 +2,15 @@ from typing import List, Any
 
 import streamlit as st
 
-from digiprod_gen.backend.browser.crawling.proxies import get_random_private_proxy
-from digiprod_gen.backend_api.browser.crawling.utils import get_random_headers
+from digiprod_gen.backend_api.browser.crawling.utils.utils_mba import get_random_headers
 from digiprod_gen.backend_api.browser.selenium_fns import SeleniumBrowser
 from digiprod_gen.backend_api.models.mba import CrawlingMBARequest, MBAMarketplaceDomain, MBAProductCategory
 from digiprod_gen.backend.data_classes.session import SessionState, ImageGenData, CrawlingData, MBAUploadData, \
     DigiProdGenStatus, DigitProdGenViews
 from digiprod_gen.backend.data_classes.config import DigiProdGenConfig
 from digiprod_gen.frontend.backend_caller import BackendCaller
-from digiprod_gen.backend_api.utils import request2mba_overview_url, is_debug, delete_files_in_path
+from digiprod_gen.backend_api.utils import is_debug, delete_files_in_path
+from digiprod_gen.backend_api.transform.transform_fns import request2mba_overview_url
 
 
 def creat_session_state(config: DigiProdGenConfig) -> SessionState:
